jobs/mongo-writer/requirements.txt
kafka-python==2.0.2
pymongo==4.3.3
python-dotenv==0.19.0